App

Description

Module contains methods for requesting GIOS API and mechanisms for retrieving eligible data.

Documentation

type WithAssets = Api :<|> Raw

withAssets :: Proxy WithAssets

app :: IO Application

Returns application instance.

server :: IO (Server WithAssets)

Creates instance of a server.

apiServer :: DB -> Server Api

Returns api routing for allowed requests.

data DB

Scheme of fake database

Constructors

DB (MVar (Map Int Table))

process

Arguments

:: Int
Index of table

-> DB
Instance of fake db

-> IO Table
Returns desired table

Main method for retrieving table.

getTimeDiff :: String -> IO Double

Computes difference between given time (as string) and current time.

generateTable :: Int -> IO Table

Generates new table if desired table does not exist or has expired.

generateStation :: Int -> IO Station

Returns instance of station data from GIOS API.

generateAvg :: Int -> IO [Double]

Computes average pollution value from last 24 surveys for each compound.

getSensorIds :: Int -> IO [String]

Returns indexes of sensors for given station.

getAvgFromUrl :: String -> IO Double

Returns average pollution value for given URL.

getValues :: String -> IO [Survey]

Handles requesting GIOS API for given sensor URL.

computeAvg :: [Survey] -> Double

Computes average pollution value for given array of values.

unwrapDouble :: Maybe Double -> Double

Unwraps double type.

unwrapTime :: Maybe String -> String

Unwraps time value.

getLevelName

Arguments

:: IO Station
Wrapped station

-> (Station -> Maybe Index)
Retrieving field from data Station

-> IO String
Retrieved index value name

Retrieves index value name for given field.

createTable

Arguments

:: IO Station
Instance of Station data

-> [Double]
Array of average pollution values

-> IO Table
Wrapped Table data instance

Returns new instance of Table data.